Extract core functionnalities to a plugin
