+++ npm/__tests__/unit/cocooned/plugins/core/triggers/add/builder.js
@@ -1,6 +1,6 @@
 /* global given */
 
-import { Builder } from '@notus.sh/cocooned/src/cocooned/triggers/add/builder'
+import { Builder } from '@notus.sh/cocooned/src/cocooned/plugins/core/triggers/add/builder'
 import { faker } from '@cocooned/tests/support/faker'
 
 describe('Builder', () => {
