+++ npm/src/cocooned/plugins/core/triggers/add/validator.js
